Refactored code, removing dead code.
